fix: make IRProgram an interface
