+++ src/main/java/me/mtk/torrey/frontend/ir/gen/IRProgram.java
@@ -1,79 +1,16 @@
 package me.mtk.torrey.frontend.ir.gen;
 
+import java.util.List;
 import me.mtk.torrey.frontend.ir.instructions.Quadruple;
 
-import java.util.ArrayList;
-import java.util.List;
-
-public final class IRProgram
+/**
+ * A sequence of three-address instructions, where
+ * each instruction is represented by a quadruple
+ * of the form (op, arg1, arg2, result).
+ */
+public interface IRProgram
 {
-  // The three-address instructions, each represented
-  // by a quadruple (op, arg1, arg2, result).
-  private List<Quadruple> quads;
-
-  public IRProgram()
-  {
-    quads = new ArrayList<>();
-  }
-
-  public void addQuad(Quadruple quad)
-  {
-    quads.add(quad);
-  }
-
-  public void addQuads(List<Quadruple> quads)
-  {
-    for (Quadruple quad : quads)
-    {
-      addQuad(quad);
-    }
-  }
-
-  public List<Quadruple> quads()
-  {
-    return quads;
-  }
-
-  @Override
-  public boolean equals(Object o)
-  {
-    if (o == this)
-    {
-      return true;
-    }
-
-    if (!(o instanceof IRProgram))
-    {
-      return false;
-    }
-
-    IRProgram that = (IRProgram)o;
-
-    if (this.quads.size() != that.quads.size())
-    {
-      return false;
-    }
-
-    return this.quads.equals(that.quads);
-  }
-
-  @Override
-  public int hashCode()
-  {
-    return this.quads.hashCode();
-  }
-
-  @Override
-  public String toString()
-  {
-    final StringBuilder sb = new StringBuilder();
-
-    for (int i = 0; i < quads.size(); i++)
-    {
-      sb.append(quads.get(i).toString())
-        .append(i == quads.size() - 1 ? "" : "\n");
-    }
-
-    return sb.toString();
-  }
+  void addQuad(Quadruple quad);
+  void addQuads(List<Quadruple> quads);
+  List<Quadruple> quads();
 }
